requirements.txt
python-dotenv==1.0.1
google-auth-oauthlib==1.2.1
google-auth==2.37.0
google-api-python-client==2.156.0
openai==1.58.1
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5